requirements.txt
aiogram==2.13.0
aiohttp==3.9.5
aiosignal==1.3.1
schedule~=1.2.1
python-dotenv~=1.0.1
attrs~=23.2.0
Babel~=2.14.0
certifi~=2024.2.2
frozenlist~=1.4.1
idna~=3.7
multidict~=6.0.5
python-dotenv~=1.0.1
schedule~=1.2.1
yarl~=1.9.4
